backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
idna==3.8
passlib==1.7.4
psycopg2==2.9.9
pyasn1==0.6.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6